requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
ckwrap==0.1.10
Cython==0.29.32
idna==3.4
numpy==1.24.1
pandas==1.5.2
python-blizzardapi==1.3.0
python-dateutil==2.8.2
pytz==2022.7
ratelimit==2.2.1
ratelimiter==1.2.0.post0
requests==2.26.0
six==1.16.0
urllib3==1.26.13
